requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.1
Werkzeug==2.0.3
SQLAlchemy==1.4.49